% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charity_andrews1950.R
\docType{data}
\name{charity_andrews1950}
\alias{charity_andrews1950}
\title{F. Emerson Andrews Charity dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 77 rows and 11 columns.
}
\source{
Data sources:
\itemize{
\item DATA: Andrews 1950: <https://www.worldcat.org/title/263227>
}
}
\usage{
charity_andrews1950
}
\description{
F. Emerson Andrews was a foundation consultant who wrote a book that compiled estimates of charitable contributions in the US from 1929-1949.
}
\references{
F. Emerson Andrews. *Philanthropic Giving*, 1950.
}
\keyword{andrews,}
\keyword{charitable,}
\keyword{charities,}
\keyword{charity,}
\keyword{donate}
\keyword{donation,}
\keyword{give,}
\keyword{giving,}
