% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charity_caf_international.R
\docType{data}
\name{charity_caf_international}
\alias{charity_caf_international}
\title{Charities Aid Foundation International Individual Giving dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 36 rows and 10 columns.
}
\source{
Data sources:
\itemize{
\item INFO: Charity Aid Foundation publications: <https://www.cafonline.org/about-us/publications>
\item DATA: Charities Aid Foundation 2006 International Comparisons (2001-2004, only one year per country): <https://www.cafonline.org/about-us/publications/archive/international-comparisons-of-charitable-giving>
\item DATA: Charities Aid Foundation 2016 International Analysis (2007-2014, only one year per country): <https://www.cafonline.org/about-us/publications/2016-publications/gross-domestic-philanthropy>
}
}
\usage{
charity_caf_international
}
\description{
Charities Aid Foundation has released two reports, both compiled from international surveys and tax data, which estimate charity as a % of GDP by country.
}
\references{
Charities Aid Foundation. "Gross Domestic Philanthropy", 2016. Charities Aid Foundation. "International Comparisons of Charitable Giving", 2006.
}
\keyword{caf,}
\keyword{charitable,}
\keyword{charities,}
\keyword{charity,}
\keyword{donate}
\keyword{donation,}
\keyword{give,}
\keyword{giving,}
