% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charity_geddes1937.R
\docType{data}
\name{charity_geddes1937}
\alias{charity_geddes1937}
\title{Geddes Poor Relief Charity datasets}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 30 rows and 11 columns.
}
\source{
Data sources:
\itemize{
\item DATA: Geddes 1937: <https://dspace.gipe.ac.in/xmlui/bitstream/handle/10973/39394/GIPE-021706.pdf?sequence=3&isAllowed=y>
}
}
\usage{
charity_geddes1937
}
\description{
Geddes authored a report that reviewed several public and private relief datasets.
}
\references{
Geddes. *Trends in Relief Expenditures 1910-1935*, 1937.
}
\keyword{charitable,}
\keyword{charities,}
\keyword{charity,}
\keyword{donate}
\keyword{donation,}
\keyword{geddes,}
\keyword{give,}
\keyword{giving,}
