% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charity_hsus1970.R
\docType{data}
\name{charity_hsus1970}
\alias{charity_hsus1970}
\title{HSUS Charity dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 406 rows and 9 columns.
}
\source{
Data sources:
\itemize{
\item INFO: HSUS all PDFs: <https://www.census.gov/library/publications/1975/compendia/hist_stats_colonial-1970.html>
}
}
\usage{
charity_hsus1970
}
\description{
The Census Bureau published "Historical Statistics of the United States, Colonial Times to 1970" in 1975. It includes several series on charity, created by Ralph Nelson in 1973.
}
\references{
Census Bureau, *Historical Statistics of the United States, Colonial Times to 1970*, 1975.
}
\keyword{charitable,}
\keyword{charities,}
\keyword{charity,}
\keyword{donate}
\keyword{donation,}
\keyword{give,}
\keyword{giving,}
\keyword{hsus,}
