% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charity_huntley1935.R
\docType{data}
\name{charity_huntley1935}
\alias{charity_huntley1935}
\title{Huntley Poor Relief Charity dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 80 rows and 12 columns.
}
\source{
Data sources:
\itemize{
\item INFO: Huntley 1935: <https://www.worldcat.org/title/1037100721>
\item DATA: Cambridge HSUS Series Bg567–570: <https://hsus.cambridge.org/HSUSWeb/toc/showDownloadableTable.do?id=Bg320-589>
}
}
\usage{
charity_huntley1935
}
\description{
Huntley authored a report on public and private outdoor relief in New York. Outdoor relief basically means "anything that you can hand someone": Cash, coal, food.
}
\references{
Huntley. *Financial Trends in Organized Social Work in New York City*, 1935.
}
\keyword{charitable,}
\keyword{charities,}
\keyword{charity,}
\keyword{donate}
\keyword{donation,}
\keyword{give,}
\keyword{giving,}
\keyword{huntley,}
