% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charity_jones1954.R
\docType{data}
\name{charity_jones1954}
\alias{charity_jones1954}
\title{John Price Jones Charity dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 186 rows and 10 columns.
}
\source{
Data sources:
\itemize{
\item DATA: Jones 1954: <https://www.worldcat.org/title/2934937>
}
}
\usage{
charity_jones1954
}
\description{
John Price Jones was a fundraiser who wrote a book that compiled estimates of charitable contributions in the US from 1921-1950.
}
\references{
John Prices Jones. The American giver; a review of American generosity, 1954.
}
\keyword{charitable,}
\keyword{charities,}
\keyword{charity,}
\keyword{donate}
\keyword{donation,}
\keyword{give,}
\keyword{giving,}
\keyword{jones,}
