% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.r
\docType{data}
\name{countries}
\alias{countries}
\title{Country Conversion Dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 312 rows and 16 columns.
}
\usage{
countries
}
\description{
Contains strings necessary to convert various political datasets to a common ("master") standard for country names and country codes. Each dataset has 1-2 columns. Constructed manually, no code.
}
\references{
Aidan Bissell-Siders (2020). "Country Conversion Dataset".
}
\keyword{datasets}
