% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths_palestine_gaza_moh.R
\docType{data}
\name{deaths_palestine_gaza_moh}
\alias{deaths_palestine_gaza_moh}
\title{Gaza Ministry of Health Death Toll dataset of 2023 October 27}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 6747 rows and 4 columns.
}
\source{
Data sources:
\itemize{
\item DATA: Full list in PDF format at original URL: <https://web.archive.org/web/20231027142503/https://www.moh.gov.ps/portal/wp-content/uploads/2023/10/%D8%AA%D9%82%D8%B1%D9%8A%D8%B1-%D9%86%D9%87%D8%A7%D8%A6%D9%8A-%D8%A8%D8%A7%D8%B3%D9%85%D8%A7%D8%A1-%D8%A7%D9%84%D8%B4%D9%87%D8%AF%D8%A7%D8%A1-1.pdf>
}
}
\usage{
deaths_palestine_gaza_moh
}
\description{
On 2023 October 27, the Palestinian Ministry of Health (MOH) in Gaza released a dataset of 6747 dead in Gaza, with names, IDs, gender, and age.
}
\references{
تقرير تفصيلي لضحايا العدوان اإلسرائيلي على قطاع غزة خالل الفترة )7 26- أكتوبر 2023( وزارةالصحةالفلسطينية
}
\keyword{death}
\keyword{deaths,}
\keyword{gaza,}
\keyword{health}
\keyword{israel,}
\keyword{ministry}
\keyword{of}
\keyword{palestine,}
\keyword{toll,}
