% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_mw.R
\docType{data}
\name{income_mw}
\alias{income_mw}
\title{Measuring Worth United States GDP dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1392 rows and 8 columns.
}
\source{
Data sources:
\itemize{
\item INFO: What Was the U.S. GDP Then? <https://www.measuringworth.com/datasets/usgdp/>
\item DATA: Construction of Annual GDP, 1790-1928 <https://www.measuringworth.com/datasets/usgdp/constructiongdp.php>
}
}
\usage{
income_mw
}
\description{
Measuring Worth is the project of economic historian Samuel Williamson and attempts to synthesize several datasets on US wealth into colonial times.
}
\references{
Louis Johnston and Samuel H. Williamson, "What Was the U.S. GDP Then?" MeasuringWorth, 2023
}
\keyword{domestic}
\keyword{gdp,}
\keyword{gross}
\keyword{income,}
\keyword{nominal}
\keyword{product,}
\keyword{real,}
\keyword{states,}
\keyword{united}
\keyword{usa,}
