% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsed-package.r
\docType{package}
\name{ipsed}
\alias{ipsed}
\alias{ipsed-package}
\title{International Political, Social, and Economic datasets}
\source{
\itemize{
\item Clio Infra (2014): \link{clio}
\item From Empire to Nation-State (2006): \link{fetns}
\item Maddison (2018): \link{maddison}
\item Penn World Table 9.1 (2019): \link{pwt}
\item Polity 5 (2020): \link{polity}
\item World Giving Index (2010-2019): \link{wgi}
}
}
\description{
A collection of political, social, and economic datasets formatted to allow between-country comparison. Datasets included: Clio Infra (2014), From Empire to Nation-state (2006), Maddison Project (2018), Penn World Table 9.1 (2019), Polity 5 (2020), World Inequality Database (2020)
}
\references{
See dataset pages
}
\author{
\strong{Maintainer}: Aidan Bissell-Siders \email{aidan.bissell.siders@gmail.com} (\href{https://orcid.org/0000-0002-6180-8324}{ORCID})

}
