% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iraq_body_count.r
\docType{data}
\name{iraq_body_count}
\alias{iraq_body_count}
\title{Iraq Body Count dataset}
\format{
A data table with 51607 or more rows and 13 variables. Key variables:
\describe{
\item{start_date}{date on which deaths started}
\item{end_date}{date on which deaths were reported}
\item{location}{location in Iraq}
\item{reported_minimum}{minimum number of civilians reported killed}
\item{reported_minimum}{maximum number of civilians reported killed}
}
}
\source{
Data sources:
\itemize{
\item DATA: Documented civilian deaths from violence: <https://www.iraqbodycount.org/database/>
\item INFO: A comparison of the Iraq Body Count and Wikileaks War Logs, showing that IBC is biased toward high-fatality events: <https://www.cambridge.org/core/journals/prehospital-and-disaster-medicine/article/abs/wikileaks-and-iraq-body-count-the-sum-of-parts-may-not-add-up-to-the-wholea-comparison-of-two-tallies-of-iraqi-civilian-deaths/93585802BA9EA65288321F492CA47A43>
\item INFO: A comparison of the Iraq Body Count and other sources: <https://pure.royalholloway.ac.uk/portal/files/32733064/Submitted_to_Casualties_Book.pdf>
}
}
\usage{
iraq_body_count
}
\description{
The Iraq Body Count is a passive surveillance count of civilian deaths in Iraq since 2003. Passive surveillance guarantees that the IBC is an undercount: only deaths corroborated by public reports (such as media, hospital, or gov't reports) are included.
}
\references{
Iraq Body Count. https://www.iraqbodycount.org/
}
\keyword{count,casualty,casualties,kill,killing,dead,death,fatality,fatal,death}
\keyword{iraq,body}
\keyword{toll}
