% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_beito2000.R
\docType{data}
\name{mutual_beito2000}
\alias{mutual_beito2000}
\title{Beito fraternity payments dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 30 rows and 12 columns.
}
\source{
Data sources:
\itemize{
\item DATA: Beito 2000: <https://www.worldcat.org/title/301229610>
}
}
\usage{
mutual_beito2000
}
\description{
Beito 2000 is a book on the history of mutual aid societies. It includes data from the National Fraternity Congress of America (now the American Fraternal Alliance) on payments from fraternities.
}
\references{
Beito. *From Mutual Aid to the Welfare State: Fraternal Societies and Social Services, 1890-1967*, 2000.
}
\keyword{afa,}
\keyword{aid,}
\keyword{beito,}
\keyword{benefit,}
\keyword{fraternities,}
\keyword{fraternity,}
\keyword{insurance,}
\keyword{mutual}
\keyword{mutual,}
\keyword{mutuals,}
\keyword{nfc,}
\keyword{nfca,}
\keyword{societies}
