% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_lindert1994.R
\docType{data}
\name{social_lindert1994}
\alias{social_lindert1994}
\title{Lindert 1994 Social Spending dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 900 rows and 9 columns.
}
\source{
Data sources:
\itemize{
\item DATA: Lindert 1994: see Tables 1A-1E: <https://www.sciencedirect.com/science/article/abs/pii/S0014498384710011>
}
}
\usage{
social_lindert1994
}
\description{
Peter Lindert is an economic historian who has extensively documented the rise of the modern welfare state.
}
\references{
Lindert P, "The Rise of Social Spending, 1880-1930", 1994.
}
\keyword{education,}
\keyword{expenditure,}
\keyword{expenses,}
\keyword{government}
\keyword{lindert,}
\keyword{pensions,}
\keyword{social}
\keyword{spending,}
\keyword{unemployment,}
\keyword{welfare,}
