% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_lindert2021_how_was_life_vol_ii.R
\docType{data}
\name{social_lindert2021_how_was_life_vol_ii}
\alias{social_lindert2021_how_was_life_vol_ii}
\title{Lindert 2021 Social Spending dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 500 rows and 9 columns.
}
\source{
Data sources:
\itemize{
\item DATA: Lindert 2021: see Tables 4.2 and 4.3: <https://www.oecd-ilibrary.org/sites/3d96efc5-en/1/3/4/index.html?itemId=/content/publication/3d96efc5-en&_csp_=2c2e680562193998e9d20ed6a45a9242&itemIGO=oecd&itemContentType=book>
}
}
\usage{
social_lindert2021_how_was_life_vol_ii
}
\description{
Peter Lindert is an economic historian who has extensively documented the rise of the modern welfare state.
}
\references{
Lindert P, "Social Spending and the Welfare State", 2021, in *How Was Life? Volume II*
}
\keyword{education,}
\keyword{expenditure,}
\keyword{expenses,}
\keyword{government}
\keyword{lindert,}
\keyword{oecd,}
\keyword{pensions,}
\keyword{social}
\keyword{spending,}
\keyword{unemployment,}
\keyword{welfare,}
