% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_lindert_1960_2019.R
\docType{data}
\name{social_lindert_1960_2019}
\alias{social_lindert_1960_2019}
\title{Lindert 1960-2019 Social Spending dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4560 rows and 8 columns.
}
\source{
Data sources:
\itemize{
\item DATA: Lindert Data Garden: <https://economics.ucdavis.edu/people/fzlinder/peter-linderts-webpage/data-garden>
}
}
\usage{
social_lindert_1960_2019
}
\description{
Peter Lindert is an economic historian who has extensively documented the rise of the modern welfare state.
}
\references{
Lindert P, "Data Garden", n.d.
}
\keyword{education,}
\keyword{expenditure,}
\keyword{expenses,}
\keyword{government}
\keyword{lindert,}
\keyword{pensions,}
\keyword{social}
\keyword{spending,}
\keyword{unemployment,}
\keyword{welfare,}
