% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_oecd1985.R
\docType{data}
\name{social_oecd1985}
\alias{social_oecd1985}
\title{OECD Social Spending 1985 dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3344 rows and 8 columns.
}
\source{
Data sources:
\itemize{
\item DATA: OECD 1985: <https://archive.org/details/socialexpenditur0000unse/page/100/mode/2up>
\item DATA: OECD 1985: <https://www.oecd.org/social/soc/40836112.pdf>
}
}
\usage{
social_oecd1985
}
\description{
Since 1950, the OECD has compiled statistics on social spending of various countries (usually the OECD plus a few others). This edition, from 1985, covers 1960 to 1981. It does not break out housing spending.
}
\references{
OECD. *Social expenditure, 1960-1990 : problems of growth and control*, 1985.
}
\keyword{education,}
\keyword{expenditure,}
\keyword{expenses,}
\keyword{government}
\keyword{oecd,}
\keyword{pensions,}
\keyword{social}
\keyword{spending,}
\keyword{unemployment,}
\keyword{welfare,}
