% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_oecd_socx_agg.R
\docType{data}
\name{social_oecd_socx_agg}
\alias{social_oecd_socx_agg}
\title{OECD Social Expenditure Database dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 32602 rows and 11 columns.
}
\source{
Data sources:
\itemize{
\item INFO: OECD SOCX database: <https://www.oecd.org/social/expenditure.htm>
\item DATA: OECD SOCX_AGG dataset: <https://stats.oecd.org/Index.aspx?datasetcode=SOCX_AGG>
}
}
\usage{
social_oecd_socx_agg
}
\description{
Since 1950, the OECD has compiled statistics on social spending of various countries (usually the OECD plus a few others).
}
\references{
OECD. "Social Expenditure Database", 2023.
}
\keyword{education,}
\keyword{expenditure,}
\keyword{expenses,}
\keyword{government}
\keyword{oecd,}
\keyword{pensions,}
\keyword{social}
\keyword{spending,}
\keyword{unemployment,}
\keyword{welfare,}
